user has to be logged in to post
