+++ spec/features/users_features_spec.rb
@@ -30,14 +30,3 @@ describe 'User can sign in and out' do
   end
 end
 
-def sign_up(email: 'psmith@makers.com', password: 'password')
-  click_link 'Sign up'
-  fill_in 'Email', with: email 
-  fill_in 'Password', with: password
-  fill_in 'Password confirmation', with: password
-  click_button 'Sign up'
-end
-
-def sign_out
-  click_link 'Sign out'
-end
